pyproject.toml
[tool.poetry]
name = "deep-memory-update"
version = "0.1.0"
description = ""
authors = ["Łukasz Neumann <lukasz.neumann@pw.edu.pl>"]

[tool.poetry.dependencies]
python = "^3.7,<3.10"
black = "^20.8b1"
numpy = "^1.20"
torch = "^1.8.0"
pytorch-nlp = "^0.5.0"
pytorch-lightning = "^1.0.8"
scipy = "^1.6.2"
mlflow = "^1.15.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
parameterized = "^0.8.1"
guildai = "^0.7.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
